+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## To be released
+- Support validating state that contain non-standard JS objects (such as Immutable.js)
+
 ## v0.1.3 (November 1, 2017)
 - Changed license to MIT
 - Added type declaration file